+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [未发布]
+
 ## [0.7.0] (2017-12-23)
 
 ### 新增
@@ -63,6 +65,7 @@
 * 完成大部分 Bucket API(还剩一个 Put Bucket Lifecycle)
 
 
+[未发布]: https://github.com/mozillazg/go-cos/compare/v0.7.0...HEAD
 [0.7.0]: https://github.com/mozillazg/go-cos/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/mozillazg/go-cos/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/mozillazg/go-cos/compare/v0.4.0...v0.5.0